[CHORE] 강황증권 파일명 수정
