+++ .gitignore
@@ -42,7 +42,7 @@ src/main/resources/config/enterprise.properties
 src/main/resources/config/redis.properties
 
 #강황증권 설정 파일
-src/main/resources/config/kg-holding.properties
+src/main/resources/config/kh-holding.properties
 
 
 # 서버 설정 관련
