--- src/main/java/com/animalfarm/mlf/config/KgHoldingConfig.java
@@ -1,21 +1,0 @@
-package com.animalfarm.mlf.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-@PropertySource("classpath:config/kg-holding.properties") // 1. 파일 경로 (본인의 파일명으로 수정)
-public class KgHoldingConfig {
-	@Bean
-    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
-        return new PropertySourcesPlaceholderConfigurer();
-    }
-	
-	@Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
