+++ src/main/java/com/animalfarm/mlf/config/KhHoldingConfig.java
@@ -1,0 +1,21 @@
+package com.animalfarm.mlf.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+@PropertySource("classpath:config/kh-holding.properties") // 1. 파일 경로 (본인의 파일명으로 수정)
+public class KhHoldingConfig {
+	@Bean
+	public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
+		return new PropertySourcesPlaceholderConfigurer();
+	}
+
+	@Bean
+	public RestTemplate restTemplate() {
+		return new RestTemplate();
+	}
+}
